+++ tests/__init__.py
@@ -1,0 +1,5 @@
+
+
+"""
+Auto Link Obsidian test package.
+"""